// src/modules/user-view/user-shift-forms/TimeSelect.types.ts
/**
 * @file TimeSelect.types.ts
 * @description TimeSelect のProps型定義。
 */
import { BaseTimeProps, BaseStyles } from "../user-shift-utils/ui-interfaces";
import { ViewStyle, TextStyle } from "react-native";

export interface TimeSelectProps extends BaseTimeProps {
  startTime?: string;
  endTime?: string;
  onStartTimeChange?: (time: string) => void;
  onEndTimeChange?: (time: string) => void;
  zIndex?: number;
}

export interface TimeSelectStyles extends BaseStyles {
  timeContainer: ViewStyle;
  timeSelect: ViewStyle;
  button: ViewStyle;
  buttonText: TextStyle;
  modalOverlay: ViewStyle;
  modalContent: ViewStyle;
  optionsContainer: ViewStyle;
  scrollContainer: ViewStyle;
  optionItem: ViewStyle;
  selectedOption: ViewStyle;
  optionText: TextStyle;
  selectedOptionText: TextStyle;
}